Remove base receiver helper, rename to ScraperController
This PR also removes duplicate code for scheduling scrapers by providing a ResourceMetricScraper
wrapper for a list of MetricScrapers.
